backend/requirements.txt
flask
PyJWT
boto3